requirements.txt
numpy
pandas
python-dotenv>=0.5.1
scikit-learn
simfile
tqdm
